Fixes to properly test core

+++ Rakefile
@@ -5,25 +5,27 @@ require 'bundler/gem_tasks'
 task :default => :test
 
 Rake::TestTask.new do |task|
+  ENV['THREESCALE_CORE_INTERNAL_API'] ||= 'http://localhost:3001/internal/'
   task.test_files = FileList['test/**/*_test.rb', 'spec/**/*_spec.rb']
   task.libs = [ 'lib', File.expand_path('.') ]
   task.verbose = true
 end
 
 task :ci do
-  backend = fork do
-    ENV['RACK_ENV'] = 'development'
-    exec('3scale_backend', 'start', '-p', '3001')
+  # assume we have already set up our own backend if the env variable is set
+  if ENV['THREESCALE_CORE_INTERNAL_API'].nil?
+    backend = fork do
+      ENV['RACK_ENV'] = 'development'
+      exec('3scale_backend', 'start', '-p', '3001')
+    end
+    sleep 10
+    at_exit { Process.kill('INT', backend) }
   end
 
-  sleep 10
-
-  at_exit { Process.kill('INT', backend) }
-
+  # instruct VCR to record http requests/responses
   ENV['FULL_BUILD'] = '1'
-  ENV['THREESCALE_CORE_INTERNAL_API'] = 'http://localhost:3001/internal/'
 
-  exit system('rake', 'test')
+  Rake::Task['test'].invoke
 end
 
 ENV['gem_push'] = '0' # don't push to rubygems.org when doing rake release
